Factored BranchHelper into several modules
